HOME

HOME

DM.DataEntry — Type.

Locator for a single data entry. Uniquely defined by a set of parameters and their corresponding formatting strings. The parameter sets are divided into folder and group hierarchy.

Fields

root – the root directory of the data entry.

format_dict – the dictionary containing formattors of corresponding parameters.

value_dict – the dictionary containing valuess of corresponding parameters.

folder_entries – the list of all parameters which define the file hierarchy of the data entry.

group_entries – the list of all parameters which define the group hierarchy of the data entry.

source

DM.DataEntry — Method.

DataEntry(root, format_dict, folder_entries, group_entries)

Initialize a DataEntry type with empty values. folder_entries and group_entries are lists of their respective hierarchies.

source

DM.dm_save — Method.

dm_save(d::DataEntry, file_name::String, groups...)

Save data to entry point d with file name file_name. Data are specified by their data names and values.

source

DM.load_task_from_json — Method.

load_task_from_json(name; root="./task")

Load task from task configuration file.

source

DM.set_value! — Method.

set_value!(data_entry, value, key)

Set the value of given parameters in DataEntry type. key and value can be a single key-value pair, or tuples of corresponding pairs.

source